# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/wrap-token/tokens-sent/{paymentId}": {
      "patch": {
        "operationId": "updateToTokensSent",
        "summary": "Update transaction status to tokens sent",
        "parameters": [
          {
            "name": "paymentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateToTokensSentReqDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessDTO"
                }
              }
            }
          }
        },
        "tags": ["wrap-token"]
      }
    },
...
    "/wrap-token-transactions-m2m/set-error": {
      "patch": {
        "operationId": "setCurrentError",
        "summary": "Set error status with M2M authentication",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ErrorUpdateRequestDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessDTO"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": ["wrap-token-transactions-m2m"],
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/tokens-unwrapped-m2m/awaiting-confirmation": {
      "patch": {
        "operationId": "updateToAwaitingConfirmation",
        "summary": "Update transaction to awaiting confirmation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTokensUnwrappedStatusDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessDTO"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": ["tokens-unwrapped-m2m"],
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ]
      }
    },
    "/tokens-unwrapped-m2m/confirmed": {
      "patch": {
        "operationId": "updateToConfirmed",
        "summary": "Update transaction to confirmed",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTokensUnwrappedStatusDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessDTO"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": ["tokens-unwrapped-m2m"],
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ]
      }
    },
    "/tokens-unwrapped-m2m/init-send-tokens": {
      "patch": {
        "operationId": "updateToInitSendTokens",
        "summary": "Update transaction to init send tokens",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTokensUnwrappedStatusDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessDTO"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": ["tokens-unwrapped-m2m"],
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ]
      }
    },
    "/tokens-unwrapped-m2m/sending-tokens": {
      "patch": {
        "operationId": "updateToSendingTokens",
        "summary": "Update transaction to sending tokens",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSendingTokensDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessDTO"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": ["tokens-unwrapped-m2m"],
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ]
      }
    },
    "/tokens-unwrapped-m2m/tokens-sent": {
      "patch": {
        "operationId": "updateToTokensSent",
        "summary": "Update transaction to tokens sent",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateToTokensSentDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessDTO"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": ["tokens-unwrapped-m2m"],
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ]
      }
    },
    "/tokens-unwrapped-m2m/set-error": {
      "patch": {
        "operationId": "setCurrentError",
        "summary": "Set error on unwrapped tokens transaction",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokensUnwrappedSetErrorDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessDTO"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": ["tokens-unwrapped-m2m"],
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/tokens-unwrapped-m2m": {
      "get": {
        "operationId": "getManyBaseTokensUnwrappedM2MControllerTokensUnwrappedEntity",
        "summary": "This endpoint is protected by M2M authentication",
        "parameters": [
          {
            "name": "fields",
            "description": "Selects resource fields. <a href=\"https://github.com/nestjsx/crud/wiki/Requests#select\" target=\"_blank\">Docs</a>",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "style": "form",
            "explode": false
          },
          {
            "name": "s",
            "description": "Adds search condition. <a href=\"https://github.com/nestjsx/crud/wiki/Requests#search\" target=\"_blank\">Docs</a>",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter",
            "description": "Adds filter condition. <a href=\"https://github.com/nestjsx/crud/wiki/Requests#filter\" target=\"_blank\">Docs</a>",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "style": "form",
            "explode": true
          },
          {
            "name": "or",
            "description": "Adds OR condition. <a href=\"https://github.com/nestjsx/crud/wiki/Requests#or\" target=\"_blank\">Docs</a>",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "style": "form",
            "explode": true
          },
          {
            "name": "sort",
            "description": "Adds sort by field. <a href=\"https://github.com/nestjsx/crud/wiki/Requests#sort\" target=\"_blank\">Docs</a>",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "style": "form",
            "explode": true
          },
          {
            "name": "join",
            "description": "Adds relational resources. <a href=\"https://github.com/nestjsx/crud/wiki/Requests#join\" target=\"_blank\">Docs</a>",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "style": "form",
            "explode": true
          },
          {
            "name": "limit",
            "description": "Limit amount of resources. <a href=\"https://github.com/nestjsx/crud/wiki/Requests#limit\" target=\"_blank\">Docs</a>",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "offset",
            "description": "Offset amount of resources. <a href=\"https://github.com/nestjsx/crud/wiki/Requests#offset\" target=\"_blank\">Docs</a>",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "page",
            "description": "Page portion of resources. <a href=\"https://github.com/nestjsx/crud/wiki/Requests#page\" target=\"_blank\">Docs</a>",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "cache",
            "description": "Reset cache (if was enabled). <a href=\"https://github.com/nestjsx/crud/wiki/Requests#cache\" target=\"_blank\">Docs</a>",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get many base response",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/GetManyTokensUnwrappedEntityResponseDto"
                    },
                    {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/TokensUnwrappedEntity"
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": ["tokens-unwrapped-m2m"],
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "WXTM-Bridge API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "in": "header",
        "type": "http"
      }
    },
    "schemas": {
...
      "UserEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "auth0Id": {
            "type": "string"
          },
          "isAdmin": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": ["id", "auth0Id", "isAdmin", "createdAt", "updatedAt"]
      },
...
      "UpdateToTokensSentReqDTO": {
        "type": "object",
        "properties": {
          "debug": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
      "SuccessDTO": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          }
        },
        "required": ["success"]
      },
...
      "ErrorUpdateDTO": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "paymentId": {
            "type": "string"
          }
        },
        "required": ["error", "paymentId"]
      },
      "ErrorUpdateRequestDTO": {
        "type": "object",
        "properties": {
          "walletTransactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErrorUpdateDTO"
            }
          }
        },
        "required": ["walletTransactions"]
      },
...
      "GetManyTokensUnwrappedEntityResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TokensUnwrappedEntity"
            }
          },
          "count": {
            "type": "number"
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "pageCount": {
            "type": "number"
          }
        },
        "required": ["data", "count", "total", "page", "pageCount"]
      },
      "TokensUnwrappedAuditEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "paymentId": {
            "type": "string"
          },
          "fromStatus": {
            "type": "string",
            "enum": [
              "created",
              "created_unprocessable",
              "awaiting_confirmation",
              "awaiting_confirmation_unprocessable",
              "confirmed",
              "confirmed_awaiting_approval",
              "confirmed_unprocessable",
              "init_send_tokens",
              "sending_tokens",
              "sending_tokens_unprocessable",
              "tokens_sent"
            ]
          },
          "toStatus": {
            "type": "string",
            "enum": [
              "created",
              "created_unprocessable",
              "awaiting_confirmation",
              "awaiting_confirmation_unprocessable",
              "confirmed",
              "confirmed_awaiting_approval",
              "confirmed_unprocessable",
              "init_send_tokens",
              "sending_tokens",
              "sending_tokens_unprocessable",
              "tokens_sent"
            ]
          },
          "note": {
            "type": "object"
          },
          "transaction": {
            "$ref": "#/components/schemas/TokensUnwrappedEntity"
          },
          "transactionId": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "paymentId",
          "transaction",
          "transactionId",
          "createdAt",
          "updatedAt"
        ]
      },
      "TokensUnwrappedEntity": {
        "type": "object",
        "properties": {
          "error": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            },
            "default": []
          },
          "id": {
            "type": "number"
          },
          "paymentId": {
            "type": "string"
          },
          "subgraphId": {
            "type": "string"
          },
          "nonce": {
            "type": "number"
          },
          "signature": {
            "type": "string"
          },
          "contractAddress": {
            "type": "string"
          },
          "from": {
            "type": "string"
          },
          "targetTariAddress": {
            "type": "string"
          },
          "amount": {
            "type": "string"
          },
          "feePercentageBps": {
            "type": "number"
          },
          "feeAmount": {
            "type": "string"
          },
          "amountAfterFee": {
            "type": "string"
          },
          "blockHash": {
            "type": "string"
          },
          "blockNumber": {
            "type": "number"
          },
          "blockTimestamp": {
            "format": "date-time",
            "type": "string"
          },
          "transactionHash": {
            "type": "string"
          },
          "status": {
            "enum": [
              "created",
              "created_unprocessable",
              "awaiting_confirmation",
              "awaiting_confirmation_unprocessable",
              "confirmed",
              "confirmed_awaiting_approval",
              "confirmed_unprocessable",
              "init_send_tokens",
              "sending_tokens",
              "sending_tokens_unprocessable",
              "tokens_sent"
            ],
            "type": "string"
          },
          "isErrorNotificationSent": {
            "type": "boolean"
          },
          "temporaryTransactionId": {
            "type": "string"
          },
          "tariTxTimestamp": {
            "type": "number"
          },
          "tariBlockHeight": {
            "type": "number"
          },
          "tariPaymentReference": {
            "type": "string"
          },
          "audits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TokensUnwrappedAuditEntity"
            }
          },
          "approvingUser": {
            "$ref": "#/components/schemas/UserEntity"
          },
          "approvingUserId": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "error",
          "id",
          "paymentId",
          "subgraphId",
          "nonce",
          "signature",
          "contractAddress",
          "from",
          "targetTariAddress",
          "amount",
          "feePercentageBps",
          "feeAmount",
          "amountAfterFee",
          "blockHash",
          "blockNumber",
          "blockTimestamp",
          "transactionHash",
          "status",
          "isErrorNotificationSent",
          "audits",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "UpdateTokensUnwrappedStatusDTO": {
        "type": "object",
        "properties": {
          "paymentId": {
            "type": "string"
          }
        },
        "required": ["paymentId"]
      },
      "UpdateSendingTokensDTO": {
        "type": "object",
        "properties": {
          "paymentId": {
            "type": "string"
          },
          "temporaryTransactionId": {
            "type": "string"
          }
        },
        "required": ["paymentId", "temporaryTransactionId"]
      },
      "UpdateToTokensSentDTO": {
        "type": "object",
        "properties": {
          "paymentId": {
            "type": "string"
          },
          "tariTxTimestamp": {
            "type": "number"
          },
          "tariBlockHeight": {
            "type": "number"
          },
          "tariPaymentReference": {
            "type": "string"
          }
        },
        "required": [
          "paymentId",
          "tariTxTimestamp",
          "tariBlockHeight",
          "tariPaymentReference"
        ]
      },
      "TokensUnwrappedSetErrorDTO": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "paymentId": {
            "type": "string"
          }
        },
        "required": ["error", "paymentId"]
      },
...
    }
  }
}